Move dcc encryption and signing to an own class.
